* Integrate new routing tree with router (which fixed ControllerTest#test_generated_routes_match_actions) * Replace Route with RouteNode * Rename WildcardRouteNode to WildcardRoute * Change Router::Tree#search to return action from found route * Fix require for test/router/tree_test.rb
