+++ test/router/route_test.rb
@@ -1,9 +1,9 @@
 require_relative "../helper"
 
 module Router
-  class RouteNodeTest < MiniTest::Unit::TestCase
+  class RouteTest < MiniTest::Unit::TestCase
     def setup
-      @node = Harbor::Router::RouteNode.new
+      @node = Harbor::Router::Route.new
       @node.insert(:index, ['posts'])
     end
 
@@ -80,7 +80,7 @@ module Router
       @node.insert(:id, ['posts', ':id'])
       @node.insert(:recent, ['posts', 'recent'])
 
-      assert_kind_of Harbor::Router::WildcardRouteNode, @node.match
+      assert_kind_of Harbor::Router::WildcardRoute, @node.match
       refute_nil @node.match.trees
     end
 
@@ -88,7 +88,7 @@ module Router
       @node.insert(:recent, ['posts', 'recent'])
       @node.insert(:id, ['posts', ':id'])
 
-      assert_kind_of Harbor::Router::WildcardRouteNode, @node.match
+      assert_kind_of Harbor::Router::WildcardRoute, @node.match
       refute_nil @node.match.trees
     end
